Move validation models to models directory
