+++ src/components/sidebar/linkView/LinkView.tsx
@@ -11,8 +11,8 @@ import ViewFormBase from '~/components/shared/inheritedComponents/ViewFormBase';
 import Loader from '~/components/shared/loader/Loader';
 import LinkService from '~/services/linkService';
 import nodeTypeCodeList from '~/codeLists/nodeTypeCodeList';
-import linkValidationModel from '~/validation/models/linkValidationModel';
 import SubSites from '~/routing/subSites';
+import linkValidationModel from '~/models/validationModels/linkValidationModel';
 import directionCodeList from '~/codeLists/directionCodeList';
 import { DialogStore } from '~/stores/dialogStore';
 import routeBuilder from '~/routing/routeBuilder';
