+++ src/components/sidebar/routePathView/routePathInfoTab/RoutePathForm.tsx
@@ -6,9 +6,9 @@ import { IRoutePath } from '~/models';
 import { RoutePathStore } from '~/stores/routePathStore';
 import booleanCodeList from '~/codeLists/booleanCodeList';
 import routeBuilder from '~/routing/routeBuilder';
+import routePathValidationModel from '~/models/validationModels/routePathValidationModel';
 import SubSites from '~/routing/subSites';
 import navigator from '~/routing/navigator';
-import routePathValidationModel from '~/validation/models/routePathValidationModel';
 import { IValidationResult } from '~/validation/FormValidator';
 import InputContainer from '../../InputContainer';
 import TextContainer from '../../TextContainer';
