+++ src/models/validationModels/linkValidationModel.ts
